refactor(video): reclassify VideoPlayer organism → molecule
VideoPlayer wraps a single native <video> with HLS attachment — a simple
functional unit, not a complex page section. Move it under components/
molecules, update the barrels, and import it from ../molecules in
PageTemplate.

Refs #24
