+++ components/molecules/index.ts
@@ -10,6 +10,11 @@
 export { Author } from './Author'
 export type { AuthorProps } from './Author'
 
+// VideoPlayer (shared HLS player)
+export { VideoPlayer } from './VideoPlayer'
+export type { VideoPlayerProps } from './VideoPlayer'
+export type { VideoSubtitleCue } from './VideoPlayer'
+
 // Content
 export { Column } from './Column'
 export type { ColumnProps } from './Column'
